+++ lists/tests/test_view.py
@@ -69,33 +69,33 @@ class HomePageTest(TestCase):
 	# 	self.assertIn('itemey 2', response.content.decode())
 
 # class ItemModelTest(TestCase):
-class ListAndItemModelsTest(TestCase):
-	def test_saving_and_retrieving_items(self):
-		list_ = List()
-		list_.save()
+# class ListAndItemModelsTest(TestCase):
+# 	def test_saving_and_retrieving_items(self):
+# 		list_ = List()
+# 		list_.save()
 
-		first_item = Item()
-		first_item.text = 'The first (ever) list item'
-		first_item.list = list_
-		first_item.save()
+# 		first_item = Item()
+# 		first_item.text = 'The first (ever) list item'
+# 		first_item.list = list_
+# 		first_item.save()
 
-		second_item = Item()
-		second_item.text = 'Item the second'
-		second_item.list = list_
-		second_item.save()
+# 		second_item = Item()
+# 		second_item.text = 'Item the second'
+# 		second_item.list = list_
+# 		second_item.save()
 
-		saved_list = List.objects.first()
-		self.assertEqual(saved_list, list_)
+# 		saved_list = List.objects.first()
+# 		self.assertEqual(saved_list, list_)
 
-		saved_items = Item.objects.all()
-		self.assertEqual(saved_items.count(), 2)
+# 		saved_items = Item.objects.all()
+# 		self.assertEqual(saved_items.count(), 2)
 
-		first_saved_item = saved_items[0]
-		second_saved_item = saved_items[1]
-		self.assertEqual(first_saved_item.text, 'The first (ever) list item')
-		self.assertEqual(first_saved_item.list, list_)
-		self.assertEqual(second_saved_item.text, 'Item the second')
-		self.assertEqual(second_saved_item.list, list_)
+# 		first_saved_item = saved_items[0]
+# 		second_saved_item = saved_items[1]
+# 		self.assertEqual(first_saved_item.text, 'The first (ever) list item')
+# 		self.assertEqual(first_saved_item.list, list_)
+# 		self.assertEqual(second_saved_item.text, 'Item the second')
+# 		self.assertEqual(second_saved_item.list, list_)
 
 class ListViewTest(TestCase):
 	def test_uses_list_template(self):
@@ -138,6 +138,7 @@ class ListViewTest(TestCase):
 
 		response = self.client.get(f'/lists/{correct_list.id}/')
 		self.assertEqual(response.context['list'], correct_list)
+
 
 class NewListTest(TestCase):
 	def test_can_save_a_POST_request(self):
